# api.doc.json
{
    "openapi":"3.0.1",
    "info": {
        "title": "NodeTodo",
        "description":"Node/Typescript API project starter pack",
        "contact": {
            "name": "simonbassey",
            "email":"simon.dev.bassey@gmail.com",
            "url":"https://github.com/simonbassey"
        },
        "licences": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes":["http"],
    "host": "http://localhost:3000",
    "basePath":"http://localhost:3000/api",
    "paths":{
        "/api/todos": {
            "summary": "Todos",
            "get": {
                "summary":"Get Todo List",
                "description": "Returns a list of todos",
                "responses": {
                    "200": {
                        "description": "A list of available todos",
                        "content": {
                            "application/json": {
                              "schema": {
                                "type": "array",
                                "items": {"$ref": "#/components/schemas/TodoItem"}
                              }
                            }
                          }
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "get":{
                "summary": "Gets a Todo",
                "description": "Gets the details of a todo by the provided Id",
                "parameters": {
                    "name": "id",
                    "in": "path",
                    "required": "true",
                    "description": "The Id of the todo to update",
                    "type": "string"
                },
                "responses":{
                    "200": {
                        "description": "A Todo documents",
                        "content": {
                            "application/json": {
                              "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/TodoItem"
                              }
                            }
                        }
                    }
                }
            }
        },
        "/api/todos/create": {
            "post": {
                "summary": "Create a new Todo",
                "description": "Creates and returns a new Todo list",
                "requestBody": {
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/TodoCreateRequest"
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TodoItem": {
                "type":"object",
                "properties": {
                    "completed": {
                        "type": "boolean"
                    },
                    "_id": { "type": "string"},
                    "title": { "type": "string"},
                    "description": { "type": "string"},
                    "userId": { "type": "string"},
                    "createdDate": { "type": "date"},
                    "lastUpdated": { "type": "date"}
                }
            },
            "TodoCreateRequest": {
                "type":"object",
                "properties": {
                    "status": {"type": "boolean"},
                    "title": { "type": "string"},
                    "description": { "type": "string"}
                }
            }
        }
    }
}